refactor : 가계부 데이터 집계 로직 분리
